vectorstore-in-memory/pyproject.toml
[project]
name = "vectorstore-in-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.13",
    "langchainhub>=0.1.21",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
]
